Remove the private worker refcell (#223)
It should be possible to drop a private bridge from within its handler as it is in a public bridge.

When using private workers, if I attempted to drop the bridge within the handler of the bridge, then an interior mutability issue would arise. This is because a refcell was being used and the handler ref is borrowed while then subsequently dropping; dropping also attempts a mutable borrow and then panics.

The change here is to do away with the refcell. Instead, the bridge callback is owned by the bridge's handler. We also take care to set the handler to an empty closure as the first part of dropping the bridge.